Refactor cpp side code
# Conflicts:
#	src/gfx/imgui/imgui.cpp
#	src/gfx/tests/sandbox.cpp
