+++ src/gfx/gfx_wgpu.hpp
@@ -22,6 +22,7 @@
 #define wgpuQueueRelease(...)
 #define wgpuAdapterRelease(...) wgpuAdapterDrop(__VA_ARGS__)
 #define wgpuSurfaceRelease(...) wgpuSurfaceDrop(__VA_ARGS__)
+#define wgpuInstanceRelease(...) wgpuInstanceDrop(__VA_ARGS__)
 
 extern "C" {
 #include <wgpu.h>
